Remove CoreUtils from Server project, was identical to Common one. Add a player count property.
